Member_management_system-main/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
SQLAlchemy==2.0.36
python-dotenv==1.0.0
pydantic-settings==2.1.0
